Draw rays through the scene to model ray tracing
The math isn't exactly right, but it's very close at least for the first
few lines. If you're in testing mode, it will draw a point instead of the
ray so that you're able to see it from the camera. This point doesn't line
up directly with the ray for some reason, but it's very close.